// src/models/content.model.ts
export interface pharmaContentAsset {
  attributes: any;
  createTime: string;
  createdBy: string;
  id: string;
  instance: string;
  isActive: boolean;
  lastModifiedTime: string;
  modifiedBy: string;
  resourceType: string;
  sharedTime?: string;
  name: string;
  type: string;
  lang: string;
  _services_client?: any;
}